Write image digest generated by oci_image out to src code
The plan is to write out this digest to a file within a helm chart which it will be used to reference an image from the registry.

Done through write_digest_from_oci custom rule. There is functionality in the root BUILD.bazel file which copies files generated from targets into a destination the src directory.